OnTapGesture in Calendar is now working.
The OnTapGesture in the cal.swift file now changes two variables.  the first is the showingAlert variable and the second is the dateSelection variable.  these State variables allow the Alert box to be displayed.  The Alert box houses the selected date and calculates 40 days past that date and displays it as the earliest closing date.  I also added the glow function in the HomeView.swift file.  this adds a glowing edge to shapes.  I changed the look and feel of the app slightly, looking for feedback on it.

The onTapGesture still does not turn the background of a specific date a different color, but it does give the user at least a little information.